+++ src/main/java/com/cyosp/ids/model/User.java
@@ -1,9 +1,13 @@
 package com.cyosp.ids.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.Size;
 
 @Data
 @Builder
@@ -17,7 +21,20 @@ public class User {
 
     private String email;
 
+    @Size(min = PASSWORD_MIN_LENGTH, max = PASSWORD_MAX_LENGTH)
     private String password;
+
+    @JsonIgnore
+    public String getPassword() {
+        return password;
+    }
+
+    @JsonProperty
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    private String hashedPassword;
 
     private Role role;
 }