+++ src/main/java/com/cyosp/ids/repository/UserRepository.java
@@ -3,29 +3,37 @@ package com.cyosp.ids.repository;
 import com.cyosp.ids.model.User;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
+import javax.validation.ConstraintViolation;
+import javax.validation.ValidationException;
+import javax.validation.Validator;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import static com.cyosp.ids.configuration.IdsConfiguration.DATA_DIRECTORY_PATH;
+import static java.util.Objects.isNull;
+import static java.util.stream.Collectors.joining;
+import static org.springframework.beans.BeanUtils.copyProperties;
 
 @Slf4j
 @Component
+@RequiredArgsConstructor
 public class UserRepository {
     private final File usersFile = new File(DATA_DIRECTORY_PATH + "ids.users.json");
 
     private final ObjectMapper objectMapper;
 
-    private List<User> users;
+    private final Validator validator;
 
-    UserRepository(ObjectMapper objectMapper) {
-        this.objectMapper = objectMapper;
-    }
+    private List<User> users;
 
     @PostConstruct
     public void init() {
@@ -33,6 +41,7 @@ public class UserRepository {
             try {
                 users = objectMapper.readValue(usersFile, new TypeReference<>() {
                 });
+                setHashedPasswords();
             } catch (Exception e) {
                 throw new RuntimeException("Fail to load users from: " + usersFile.getAbsolutePath() + "\n" + e.getMessage());
             }
@@ -42,23 +51,58 @@ public class UserRepository {
         }
     }
 
+    private void setHashedPasswords() {
+        AtomicBoolean needSaveUsers = new AtomicBoolean(false);
+        users.forEach(user -> {
+            if (isNull(user.getHashedPassword())) {
+                user.setHashedPassword(user.getPassword());
+                user.setPassword(null);
+                needSaveUsers.set(true);
+            }
+        });
+        if (needSaveUsers.get()) {
+            saveUsers();
+        }
+    }
+
     public User save(User user) {
+        validate(user);
+        users.removeIf(u -> u.getId().equals(user.getId()));
         users.add(user);
+        saveUsers();
+        return user;
+    }
+
+    private void saveUsers() {
         try {
             objectMapper.writeValue(usersFile, users);
         } catch (Exception e) {
             throw new RuntimeException("Fail to save users into: " + usersFile.getAbsolutePath() + "\n" + e.getMessage());
         }
-        return user;
     }
 
     public List<User> findAll() {
         return users;
     }
 
-    public Optional<User> findByEmail(String email) {
+    public User getByEmail(String email) {
         return users.stream()
                 .filter(user -> email.equals(user.getEmail()))
-                .findFirst();
+                .findFirst()
+                .map(foundUser -> {
+                    User user = new User();
+                    copyProperties(foundUser, user);
+                    return user;
+                })
+                .orElseThrow(() -> new UsernameNotFoundException("Email not found: " + email));
+    }
+
+    private void validate(User user) {
+        Set<ConstraintViolation<User>> constraintViolations = validator.validate(user);
+        if (!constraintViolations.isEmpty()) {
+            throw new ValidationException(constraintViolations.stream()
+                    .map(ConstraintViolation::getMessage)
+                    .collect(joining(", ")));
+        }
     }
 }